Remove logging to eliminate console dependency
Since view.xql imports indexes.xqm, we effectively had a dependency on the console module for all views. This may not fix views that actually require the console on tomcat (see #9), namely, the remote console pane, but my hope is that this allows more users to access monex panes like monitoring.