Make sure Monex can accurately report new range index fields.
If the new range index is used, `indexes:show-index-keys()` checks the
output of `range:index-keys-for-field()` for a mismatch in the number of
terms, and more terms than asked for in `$indexes:max-number-returned`.
If the range function is acting on a document-by-document basis, Monex
cleans up the output for accuracy across the collection.

Also made it easier to tell what's going on when processing `$sorted-keys`.